[MIG] sale_product_pack: Migration to 19.0

+++ sale_product_pack/models/sale_order.py
@@ -1,6 +1,6 @@
 # Copyright 2019 Tecnativa - Ernesto Tejeda
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from odoo import _, api, models
+from odoo import api, models
 from odoo.exceptions import UserError
 
 
@@ -9,11 +9,13 @@ class SaleOrder(models.Model):
 
     def copy(self, default=None):
         sale_copy = super().copy(default)
-        # we unlink pack lines that should not be copied
-        pack_copied_lines = sale_copy.order_line.filtered(
-            lambda line: line.pack_parent_line_id.order_id == self
-        )
-        pack_copied_lines.unlink()
+        for record in self:
+            # we unlink pack lines that should not be copied
+            pack_copied_lines = sale_copy.order_line.filtered(
+                lambda x, order=record: x.pack_parent_line_id.order_id == order
+            )
+            if pack_copied_lines:
+                pack_copied_lines.unlink()
         return sale_copy
 
     @api.onchange("order_line")
@@ -31,7 +33,7 @@ class SaleOrder(models.Model):
             and not x.pack_parent_line_id.product_id.pack_modifiable
         ):
             raise UserError(
-                _(
+                self.env._(
                     "You cannot delete this line because is part of a pack in"
                     " this sale order. In order to delete this line you need to"
                     " delete the pack itself"
